requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
flask>=2.3.0
pillow>=10.0.0
requests>=2.31.0
flask-cors>=4.0.0
gunicorn>=21.0.0